requirements.txt
pdfplumber
python-dotenv
pydantic
pandas
numpy
scikit-learn
plotly
langchain
langchain-core
langchain-groq
langchain-google-genai
fastapi
uvicorn
python-multipart
pypdf
langchain-community
sentence-transformers
faiss-cpu
tiktoken
langchain-huggingface
langchain-openai
